Drop anonymous classes
